fixing for deleteVpnProfile when profile already removed

+++ src/frontends/android/app/src/main/java/org/strongswan/android/data/VpnProfileDataSource.java
@@ -220,6 +220,9 @@ public class VpnProfileDataSource
 	 */
 	public boolean deleteVpnProfile(VpnProfile profile)
 	{
+		if(profile == null){
+			return true;
+		}
 		String name = profile.getName();
 		return mDatabase.delete(TABLE_VPNPROFILE, KEY_NAME + " = \"" + name + "\"", null) > 0;
 	}
